Move settings to the OpenAI integration
